chore(build): prerelease wip

+++ install/zsh-plugins.sh
@@ -39,9 +39,11 @@ zsh-plugins::install()
 
 	mkdir -p "$ZSH_CUSTOM_PLUGINS"
 
-	log::file "Adding PPA Repository for Diff-So-Fancy"
-	sudo add-apt-repository ppa:aos1/diff-so-fancy
-	sudo apt update
+	if ! command -v diff-so-fancy > /dev/null; then
+		log::file "Adding PPA Repository for Diff-So-Fancy"
+		sudo add-apt-repository ppa:aos1/diff-so-fancy
+		sudo apt update
+	fi
 
 	echo
 	echo "--------------------------------------------------------------------"
@@ -49,21 +51,29 @@ zsh-plugins::install()
 	echo "--------------------------------------------------------------------"
 	echo
 
-	log::file "Installing Diff-So-Fancy Application"
-	sudo apt install -y diff-so-fancy
+	if ! command -v diff-so-fancy > /dev/null; then
+		log::file "Installing Diff-So-Fancy Application"
+		sudo apt install -y diff-so-fancy
+	fi
 
-	log::file "Installing Git-Flow Application"
-	sudo apt install -y git-flow
+	if ! command -v git-flow > /dev/null; then
+		log::file "Installing Git-Flow Application"
+		sudo apt install -y git-flow
+	fi
 
-	log::file "Installing FZF Fuzzy Finder Application"
-	git clone git@github.com:junegunn/fzf "$HOME"/.fzf
-	chmod 0755 "$HOME"/.fzf/install
-	cd "$HOME/.fzf" || return 1
-	./install
-	cd - || return 1
+	if [[ ! -d "$HOME/.fzf" ]]; then
+		log::file "Installing FZF Fuzzy Finder Application"
+		git clone git@github.com:junegunn/fzf "$HOME"/.fzf
+		chmod 0755 "$HOME"/.fzf/install
+		cd "$HOME/.fzf" || return 1
+		./install
+		cd - || return 1
+	fi
 
-	log::file "Installing Navi Application"
-	bash <(curl -sL https://raw.githubusercontent.com/denisidoro/navi/master/scripts/install)
+	if [[ ! -f "$HOME/.cargo/bin/navi" ]]; then
+		log::file "Installing Navi Application"
+		bash <(curl -sL https://raw.githubusercontent.com/denisidoro/navi/master/scripts/install)
+	fi
 
 	cd "$XDG_DOWNLOAD_DIR" || return 1
 
@@ -73,11 +83,15 @@ zsh-plugins::install()
 	echo "--------------------------------------------------------------------"
 	echo
 
-	log::file "Installing Git-Flow-Completion Plugin"
-	git clone git@github.com:bobthecow/git-flow-completion "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/git-flow-completion
+	if [[ ! -d "$ZSH_CUSTOM/plugins/git-flow-completion" ]]; then
+		log::file "Installing Git-Flow-Completion Plugin"
+		git clone git@github.com:bobthecow/git-flow-completion "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/git-flow-completion
+	fi
 
-	log::file "Installing ZSH Completions Plugin"
-	git clone git@github.com:zsh-users/zsh-completions "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/zsh-completions
+	if [[ ! -d "$ZSH_CUSTOM/plugins/zsh-completions" ]]; then
+		log::file "Installing ZSH Completions Plugin"
+		git clone git@github.com:zsh-users/zsh-completions "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/zsh-completions
+	fi
 
 	echo
 	echo "--------------------------------------------------------------------"
@@ -85,24 +99,36 @@ zsh-plugins::install()
 	echo "--------------------------------------------------------------------"
 	echo
 
-	log::file "Installing Bash-Insulter Plugin"
-	git clone git@github.com:hkbakke/bash-insulter
-	sudo cp bash-insulter/src/bash.command-not-found /etc/
+	if [[ ! -f /etc/bash.command-not-found ]]; then
+		log::file "Installing Bash-Insulter Plugin"
+		git clone git@github.com:hkbakke/bash-insulter
+		sudo cp bash-insulter/src/bash.command-not-found /etc/
+	fi
 
-	log::file "Installing Diff-So-Fancy Plugin"
-	git clone git@github.com:z-shell/zsh-diff-so-fancy "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/zsh-diff-so-fancy
+	if [[ ! -d "$ZSH_CUSTOM/plugins/" ]]; then
+		log::file "Installing Diff-So-Fancy Plugin"
+		git clone git@github.com:z-shell/zsh-diff-so-fancy "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/zsh-diff-so-fancy
+	fi
 
-	log::file "Installing enhan/cd Plugin"
-	git clone git@github.com:b4b4r07/enhancd "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/enhancd
+	if [[ ! -d "$ZSH_CUSTOM/plugins/" ]]; then
+		log::file "Installing enhan/cd Plugin"
+		git clone git@github.com:b4b4r07/enhancd "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/enhancd
+	fi
 
-	log::file "Installing FZF ZSH Plugin"
-	git clone git@github.com:unixorn/fzf-zsh-plugin "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/fzf-zsh-plugin
+	if [[ ! -d "$ZSH_CUSTOM/plugins/" ]]; then
+		log::file "Installing FZF ZSH Plugin"
+		git clone git@github.com:unixorn/fzf-zsh-plugin "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/fzf-zsh-plugin
+	fi
 
-	log::file "Installing ZSH AutoSuggestions Plugin"
-	git clone https://github.com/zsh-users/zsh-autosuggestions "${ZSH_CUSTOM:-"$HOME"/.oh-my-zsh/custom}"/plugins/zsh-autosuggestions
+	if [[ ! -d "$ZSH_CUSTOM/plugins/" ]]; then
+		log::file "Installing ZSH AutoSuggestions Plugin"
+		git clone https://github.com/zsh-users/zsh-autosuggestions "${ZSH_CUSTOM:-"$HOME"/.oh-my-zsh/custom}"/plugins/zsh-autosuggestions
+	fi
 
-	log::file "Installing ZSH Syntax Highlighting Plugin"
-	git clone git@github.com:zsh-users/zsh-syntax-highlighting "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/zsh-syntax-highlighting
+	if [[ ! -d "$ZSH_CUSTOM/plugins/" ]]; then
+		log::file "Installing ZSH Syntax Highlighting Plugin"
+		git clone git@github.com:zsh-users/zsh-syntax-highlighting "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"/plugins/zsh-syntax-highlighting
+	fi
 
 	cd - || return 1
